Update SensoAPIInterface to avoid duplicate codes
