+++ spring-boot/src/main/java/packages/connectouterentity/ConnectSensoScoreAPI.java
@@ -93,24 +93,11 @@ public class ConnectSensoScoreAPI implements SensoAPIInterface{
      */
     private HashMap<String, Object> CallSensoAPI(String inputJson) throws IOException, InterruptedException{
 
-        String postEndpoint = "https://auto-loan-api.senso.ai/score";
+        String postEndpoint = System.getenv("SENSO_API_SCORE_URL");
 
         // Build and send a POST request to senso endpoint
-        var request = HttpRequest.newBuilder()
-                .uri(URI.create(postEndpoint))
-                .header("Content-Type", "application/json")
-                .header("x-api-key", System.getenv("SENSO_API_KEY"))
-                .POST(HttpRequest.BodyPublishers.ofString(inputJson))
-                .build();
-
-        var client = HttpClient.newHttpClient();
-
-        // Get response from senso api
-        var response = client.send(request, HttpResponse.BodyHandlers.ofString());
-
-        Gson gson = new Gson();
-        Type mapType = new TypeToken<Map<Object, Object>>(){}.getType();
-        return gson.fromJson(response.body(), mapType);
+        HttpRequest request = buildPostRequest(inputJson, postEndpoint);
+        return getPostResponse(request);
     }
 
     /**
