+++ spring-boot/src/main/java/packages/returninfo/SignInEndpointHandler.java
@@ -25,7 +25,7 @@ public class SignInEndpointHandler {
      * @param reqBody - the body of the http request from the client
      * @return An HTTP Response back to the client.
      */
-    @CrossOrigin(origins = "http://ec2-18-118-163-255.us-east-2.compute.amazonaws.com:8080")
+    @CrossOrigin
     @PostMapping("/signin")
     public String httpResponseSignIn(@RequestBody() String reqBody) {
         System.out.println("==== POST Request Received ====");
